requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.1
django-filter
psycopg2-binary==2.9.2
pytz==2021.3
sqlparse==0.4.2
